requirements.txt
bitmath
eliot
oslo.config
psycopg2
pywbem
PyYAML
redis
six
twisted
zope.interface